+++ src/lib/selection-schemes/linear.js
@@ -4,6 +4,8 @@ import { isBefore } from 'date-fns/isBefore'
 
 import * as dateUtils from '../date-utils.js'
 
+const isDate = (time: any): boolean => time instanceof Date
+
 const linear = (selectionStart: ?Date, selectionEnd: ?Date, dateList: Array<Array<Date>>): Array<Date> => {
   let selected: Array<Date> = []
   if (selectionEnd == null) {
@@ -16,6 +18,7 @@ const linear = (selectionStart: ?Date, selectionEnd: ?Date, dateList: Array<Arra
           ? acc.concat(
               dayOfTimes.filter(
                 (t) =>
+                  isDate(t) &&
                   selectionStart &&
                   selectionEnd &&
                   dateUtils.dateHourIsBetween(